+++ README.md
@@ -53,16 +53,14 @@ open .env
 Run some code:
 
 ```js
-import { categories } from '../../ws-client-endpoints';
-import { init } from '../../ws-config';
-import { CategoryWritable } from '../types/ws-entities/categories.type';
-...
-const categoryData: CategoryWritable = {
-  ...
-}
-init('http://localhost:8080', 'GENERATE_A_COMPLEX_WSKEY_WITH_32');
-const category: Category = await categories.create(categoryData);
-console.log(`Category (id=${category.id} created at ${category.date_add}`,);
+const client = new BaseClient({
+  baseURl: process.env.BASE_URL, // URL of your PrestaShop
+  wsKey: process.env.WS_KEY, // Key to connect to your prestashop
+});
+
+const response = await client.product.create(productData);
+console.log(response.status);
+console.log(response.data.product); // Maybe return type writable ??
 
 // Error case
 // Return an Exeption depending on what is the error.
