--- jest.config.js
@@ -1,7 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  transform: { "^.+\\.ts$": "ts-jest" },
-  silent: false,
-};
