+++ package.json
@@ -46,7 +46,6 @@
   "devDependencies": {
     "@types/xml2js": "^0.4.14",
     "@types/jsdom": "^21.1.6",
-    "@types/jest": "^29.5.10",
     "@types/node": "^20.8.4",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
     "eslint": "^8.51.0",
@@ -59,7 +58,6 @@
     "prettier": "^3.0.3",
     "prettier-plugin-organize-imports": "^3.2.4",
     "rimraf": "^5.0.5",
-    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   }
