+++ src/clients/ws-orders-client.ts
@@ -1,0 +1,8 @@
+import { Endpoint } from '../enums/endpoint.enum';
+import { BaseClient } from '../ws-base-client';
+
+export class WsOrdersClient extends BaseClient<Endpoint.orders> {
+  constructor() {
+    super(Endpoint.orders);
+  }
+}
