+++ src/clients/ws-products-client.ts
@@ -1,0 +1,8 @@
+import { Endpoint } from '../enums/endpoint.enum';
+import { BaseClient } from '../ws-base-client';
+
+export class WsProductsClient extends BaseClient<Endpoint.products> {
+  constructor() {
+    super(Endpoint.products);
+  }
+}
